+++ src/lightly_train/_transforms/transform.py
@@ -23,7 +23,7 @@ from pydantic import ConfigDict, Field
 
 from lightly_train._configs.config import PydanticConfig
 from lightly_train._configs.validate import no_auto
-from lightly_train.types import TransformInput, TransformOutput
+from lightly_train.types import ImageSizeTuple, TransformInput, TransformOutput
 
 logger = logging.getLogger(__name__)
 
@@ -206,9 +206,7 @@ class RandomCropArgs(PydanticConfig):
 
 
 class MethodTransformArgs(PydanticConfig):
-    # Strict is set to False because OmegaConf does not support parsing tuples from the
-    # CLI. Setting strict to False allows Pydantic to convert lists to tuples.
-    image_size: tuple[int, int]
+    image_size: ImageSizeTuple
     channel_drop: ChannelDropArgs | None
     num_channels: int | Literal["auto"]
     random_resize: RandomResizeArgs | None