+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fix `image_size` not tuple when training from pretrained model.
+
 ### Security
 
 ## [0.12.1] - 2025-11-13